# docs/swagger.yaml
basePath: /api
definitions:
  models.CashMovement:
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@b-resto.com
    name: Soporte API
  description: API RESTful para sistema POS de restaurantes
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: B-Resto API
  version: "1.0"
paths:
...
  /kitchen-tickets:
    get:
      consumes:
      - application/json
      description: Obtiene lista de tickets de cocina con filtros
      parameters:
      - description: Filtrar por estado
        enum:
        - pending
        - preparing
        - ready
        - delivered
        in: query
        name: status
        type: string
      - description: Filtrar por estación
        in: query
        name: station_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: array de kitchen tickets'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: mensaje'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar tickets de cocina
      tags:
      - kitchen-tickets
  /kitchen-tickets/{id}:
    get:
      consumes:
      - application/json
      description: Obtiene un ticket de cocina por ID
      parameters:
      - description: ID del ticket
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: kitchen ticket'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Kitchen ticket not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener ticket de cocina
      tags:
      - kitchen-tickets
...
  /kitchen-tickets/{id}/preparing:
    patch:
      consumes:
      - application/json
      description: Cambia el estado del ticket a preparing
      parameters:
      - description: ID del ticket
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Kitchen ticket not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Marcar ticket como preparando
      tags:
      - kitchen-tickets
...
  /kitchen-tickets/{id}/status:
    patch:
      consumes:
      - application/json
      description: Cambia el estado de un ticket de cocina
      parameters:
      - description: ID del ticket
        in: path
        name: id
        required: true
        type: integer
      - description: 'status: nuevo estado'
        in: body
        name: status
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Kitchen ticket not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Actualizar estado de ticket
      tags:
      - kitchen-tickets
  /kitchen-tickets/station/{station_id}:
    get:
      consumes:
      - application/json
      description: Obtiene todos los tickets de una estación de cocina específica
      parameters:
      - description: ID de la estación
        in: path
        name: station_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: array de kitchen tickets'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: mensaje'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener tickets por estación
      tags:
      - kitchen-tickets
...
  /pos-sessions/{session_id}/cash-movements:
    get:
      consumes:
      - application/json
      description: Obtiene lista de movimientos de efectivo por sesión
      parameters:
      - description: ID de la sesión
        in: path
        name: session_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: array de cash movements'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: mensaje'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar movimientos de efectivo
      tags:
      - cash-movements
    post:
      consumes:
      - application/json
      description: Registra un nuevo movimiento de efectivo (ingreso/egreso)
      parameters:
      - description: ID de la sesión
        in: path
        name: session_id
        required: true
        type: integer
      - description: Datos del movimiento
        in: body
        name: movement
        required: true
        schema:
          $ref: '#/definitions/models.CashMovement'
      produces:
      - application/json
      responses:
        "201":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Crear movimiento de efectivo
      tags:
      - cash-movements
  /pos-sessions/{session_id}/cash-movements/{movement_id}:
    delete:
      consumes:
      - application/json
      description: Elimina un movimiento de efectivo de una sesión de caja
      parameters:
      - description: ID de la sesión
        in: path
        name: session_id
        required: true
        type: integer
      - description: ID del movimiento
        in: path
        name: movement_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Cash movement deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Cash movement not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Eliminar movimiento de efectivo
      tags:
      - cash-movements
...
securityDefinitions:
  Bearer:
    description: Escribir "Bearer {token}" para autenticación JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"